Unlinked JPEG lib DLLs
and replaced them with source JPEG library files
Removed unused files from the JPEG library (cderror.h, cdjpeg.c,
cdjpeg.h)
Moved function ChangeOrientation from separate jpegtan files to
Direct2DRenderer
Fixed bug where the return value for GetPixelFormat was not checked
Replaced strncpy with more secure strncpy_s in jerror.h
Fixed some memory variable types to be size_t in jpeglib.h

+++ ImageViewer/jpeglib.h
@@ -855,7 +855,7 @@ struct jpeg_memory_mgr {
    * used for virtual-array buffers.)  May be changed by outer application
    * after creating the JPEG object.
    */
-  long max_memory_to_use;
+  size_t max_memory_to_use;
 
   /* Maximum allocation request accepted by alloc_large. */
   long max_alloc_chunk;
@@ -977,7 +977,7 @@ EXTERN(void) jpeg_stdio_src JPP((j_decompress_ptr cinfo, FILE * infile));
 /* Data source and destination managers: memory buffers. */
 EXTERN(void) jpeg_mem_dest JPP((j_compress_ptr cinfo,
 			       unsigned char ** outbuffer,
-			       unsigned long * outsize));
+			       size_t * outsize));
 EXTERN(void) jpeg_mem_src JPP((j_decompress_ptr cinfo,
 			      unsigned char * inbuffer,
 			      unsigned long insize));
